fixed emit-types
