+++ app/shared/types/pageSchema.ts
@@ -1,40 +1,4 @@
-import Ajv from 'ajv';
-
 import { objectIdSchema } from '#shared/types/commonSchemas.js';
-import { wrapValidator } from '#shared/tsUtils.js';
-import templates from '#api/core/v1_layer/templates/index.js';
-import { PageType } from './pageType.js';
-
-const ajv = new Ajv({ allErrors: true });
-ajv.addVocabulary(['tsType']);
-
-ajv.addKeyword({
-  keyword: 'validatePageIsNotEntityView',
-  async: true,
-  errors: true,
-  type: 'object',
-  async validate(_fields: any, page: PageType) {
-    if (page.sharedId) {
-      const templatesUsingPage = await templates.getByEntityViewPage(page.sharedId);
-
-      if (templatesUsingPage.length > 0 && !page.entityView) {
-        const templatesTitles = templatesUsingPage.map(template => template.name);
-        throw new Ajv.ValidationError([
-          {
-            keyword: 'validatePageIsNotEntityView',
-            schemaPath: '',
-            params: { keyword: 'validatePageIsEntityView', _fields },
-            message: `This page is in use by the following templates: ${templatesTitles.join(
-              ', '
-            )}`,
-            instancePath: '.pages',
-          },
-        ]);
-      }
-    }
-    return true;
-  },
-});
 
 export const PageDraftSchema = {
   $schema: 'http://json-schema.org/schema#',
@@ -196,7 +160,4 @@ export const PageEditorSchema = {
   required: ['locales'],
 };
 
-const validatePage = wrapValidator(ajv.compile(PageSchema));
-const validatePageEditor = wrapValidator(ajv.compile(PageEditorSchema));
-export { validatePage, validatePageEditor };
 export const emitSchemaTypes = true;
